requirements.txt
openai>=1.30.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
schedule>=1.2.0
